Move types/const into single file, collapse some constants down
